Add class representations of components
Each component will have its own set of rendering options and defaults.
In order to manage these, each component will be backed by a class that
determines the correct settings to use for the component. This adds
initial implementations for the first 2 components using the image with
its default width and height as an example of locals that require
defaulting for a specific component.