Allow components to support nesting
This allows components to act as containers for other components or
markup by accepting and yielding blocks in the same way as render.

There is a slight awkwardness with the current implementation. When
using `render layout: ...` it is necessary to pass in a block. Otherwise
the layout option is ignored. The related Rails code is at

https://github.com/rails/rails/blob/71c7fd101324046995d8f7e51e78475c0e37ec1a/actionview/lib/action_view/helpers/rendering_helper.rb#L30

Therefore components are strictly currently containers or not
containers. The only way for a component to be both is for it to be a
container and pass in an empty block.

There are a few ways this could be potentially be improved however at
this stage it seems best to go with a simple option and see how the
components API evolves.